refactor : 예매 완료 화면 Presenter 초기화시 예매 정보를 호출하도록 수정
- 예매 완료 화면 Ui Test 패키지 수정
