+++ app/src/androidTest/java/woowacourse/movie/view/complete/BookingCompleteActivityTest.kt
@@ -1,11 +1,10 @@
-package woowacourse.movie.view.booking
+package woowacourse.movie.view.complete
 
 import android.content.Intent
 import androidx.test.core.app.ActivityScenario
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withText
+import androidx.test.espresso.Espresso
+import androidx.test.espresso.assertion.ViewAssertions
+import androidx.test.espresso.matcher.ViewMatchers
 import org.junit.Before
 import org.junit.Test
 import woowacourse.movie.domain.model.booking.PeopleCount
@@ -14,7 +13,6 @@ import woowacourse.movie.domain.model.seat.Row
 import woowacourse.movie.domain.model.seat.Seat
 import woowacourse.movie.domain.model.ticket.Ticket
 import woowacourse.movie.fixture.fakeContext
-import woowacourse.movie.view.complete.BookingCompleteActivity
 import java.time.LocalDate
 import java.time.LocalTime
 
@@ -40,9 +38,13 @@ class BookingCompleteActivityTest {
 
     @Test
     fun `전달_받은_영화_이름_상영_시간_예매_인원_예매_가격을_출력한다`() {
-        onView(withText("해리 포터와 마법사의 돌")).check(matches(isDisplayed()))
-        onView(withText("2025.4.1 12:00")).check(matches(isDisplayed()))
-        onView(withText("일반 2명")).check(matches(isDisplayed()))
-        onView(withText("26,000원 (현장 결제)")).check(matches(isDisplayed()))
+        Espresso.onView(ViewMatchers.withText("해리 포터와 마법사의 돌"))
+            .check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
+        Espresso.onView(ViewMatchers.withText("2025.4.1 12:00"))
+            .check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
+        Espresso.onView(ViewMatchers.withText("일반 2명"))
+            .check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
+        Espresso.onView(ViewMatchers.withText("26,000원 (현장 결제)"))
+            .check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
     }
 }
